Move test setup/teardown to new in_toto_test.go
